[john] - no longer segregating users into distinct groups for each email in a multiple email blast
Grouping assumed that database increments user ids by 1. This breaks for example with ClearDB which increments by 10. Generally, grouping users before applying the list rules is problematic b/c the distribution of user characteristics that list rules are applied to is not even across user ids. We care that users don't receive more than one email within a push. There is a check for this.